CLEANUP: Dry up RubyAckedMemoryQueue and RubyAckedQueue
